% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{data.Anxiety}
\alias{data.Anxiety}
\title{Anxiety ratings by different raters}
\format{A data frame with 30 observations (psychiatric diagnoses with levels 1. Depression, 2. Personality Disorder, 3. Schizophrenia, 4. Neurosis, 5. Other) on 6 variables representing different raters.}
\source{
Artificial data
}
\usage{
data(data.Anxiety)
}
\description{
The data frame contains the anxiety ratings of 20 subjects, rated by 3 raters. Values are ranging from 1 (not anxious at all) to 6 (extremely anxious).
}
\examples{
data(data.Anxiety)
data.Anxiety
}
\keyword{datasets}

