% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{data.BFI10responses}
\alias{data.BFI10responses}
\title{BFI10 dataset}
\format{Contains 45 survey responses (in rows) to the 10-point Big5 questionnaire.}
\usage{
data(data.BFI10responses)
}
\description{
A sample dataset containing SUS questionnaire responses.
}
\examples{
data(data.BFI10responses)
questionnaire.analyse(data.BFI10responses,"BFI10")
}
\keyword{datasets}

