% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{data.BFI44responses}
\alias{data.BFI44responses}
\title{BFI44 dataset}
\format{Contains 45 survey responses (in rows) to the 44-point Big5 questionnaire.}
\usage{
data(data.BFI44responses)
}
\description{
A sample dataset containing SUS questionnaire responses.
}
\examples{
data(data.BFI44responses)
questionnaire.analyse(data.BFI44responses,"BFI44")
}
\keyword{datasets}

