% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{data.Diagnoses}
\alias{data.Diagnoses}
\title{Phsychiatric Diagnoses dataset}
\format{A data frame with 30 observations (psychiatric diagnoses with levels 1. Depression, 2. Personality Disorder, 3. Schizophrenia, 4. Neurosis, 5. Other) on 6 variables representing different raters.}
\usage{
data(data.Diagnoses)
}
\description{
Psychiatric diagnoses of n=30 patients provided by different sets of m=6 raters. Data were used by Fleiss (1971) to illustrate the computation of Kappa for m raters.
}
\examples{
data(data.Diagnoses)
interrater.reliability(data.Diagnoses[,1:2],type="categorical")
}
\references{
Fleiss, J.L. (1971). Measuring nominal scale agreement among many raters. Psychological Bulletin, 76, 378-382.
}
\keyword{datasets}

