% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{data.SUSresponses}
\alias{data.SUSresponses}
\title{SUS dataset}
\format{Contains 45 survey responses (in rows) to the 10-point SUS questionnaire.}
\usage{
data(data.SUSresponses)
}
\description{
A sample dataset containing SUS questionnaire responses.
}
\examples{
data(data.SUSresponses)
questionnaire.analyse(data.SUSresponses,"SUS")
}
\keyword{datasets}

