\name{hcitools-package}
\alias{hcitools-package}
\alias{hcitools}
\docType{package}
\title{
hcitools
}
\description{
More about what it does (maybe more than one line)
A concise (1-5 lines) description of the package
}
\details{
\tabular{ll}{
Package: \tab hcitools\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-04-03\cr
License: \tab What license is it under? Free as in beer\cr
}
An overview of how to use the package, including the most important functions
}
\author{
Vassilis Kostakos, University of Melbourne

Maintainer: Who to complain to <yourfault@somewhere.net>
The author and/or maintainer of the package
}
\references{
Literature or other references for background information.
}

\keyword{ package }

