% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interrater.reliability.R
\name{interrater.reliability}
\alias{interrater.reliability}
\title{Conduct interrater reliability tests.}
\usage{
interrater.reliability(d, type = c("categorical", "ordinal"))
}
\arguments{
\item{d}{the dataset that you want to analyse. Should contain one column per rater.}

\item{type}{indicates whether the data to be analysed is categorical or ordinal. If type is ordinal then weights are used.  Make sure that for ordinal data the factors are set in the correct order. Use \code{levels(d)} to see the factor levels of your variable}
}
\value{
The results of the test, along with details of the analysis
}
\description{
\code{interrater.reliability} will perform some basic heuristics to determine which test to run on your data.
}
\examples{
data(data.Diagnoses)
interrater.reliability(data.Diagnoses[,1:2],type="categorical")
}

