% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/questionnaire.analyse.R
\name{questionnaire.analyse}
\alias{questionnaire.analyse}
\title{Analyse data from a standardised questionnaire (SUS, BFI10, BFI44) and
output the result of the analysis.}
\usage{
questionnaire.analyse(data, name = c("SUS", "BFI10", "Big10"))
}
\arguments{
\item{data}{A data frame containing (only) the questionnaire responses.}

\item{name}{The name of the questionaire. Can be 'SUS', 'BFI44' or 'BFI10'.}
}
\value{
The analysis of the questionnaire
}
\description{
\code{questionnaires_analyse} will simply print out a standard questionnaire,
and the citation for it.
}
\details{
Function's \code{data} parameter must contain all the questions as columns
generated by \code{\link{questionnaire.show}} as columns in a data frame. The
columns must be in the same order as in the questionnaire printed
by \code{\link{questionnaire.show}}. No other columns should exist in the
\code{data} parameter. Column names do not matter.
}
\examples{
require(hcitools)
data(data.SUSresponses)
questionnaire.analyse(data.SUSresponses,'SUS')
}

