% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/questionnaire.show.R
\name{questionnaire.show}
\alias{questionnaire.show}
\title{Print a version of a standardised questionnaire}
\usage{
questionnaire.show(name = c("SUS", "BFI10", "BFI44"))
}
\arguments{
\item{name}{the name of the questionaire. Can be 'SUS', 'Big5', etc.}
}
\value{
Prints instructions, the questionnaire in plain text, and a citation.
}
\description{
\code{questionnaire.show} will simply print out a standard questionnaire,
and the citation for it
}
\examples{
questionnaire.show('SUS')
}

