% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hcitools.R
\name{questionnaire_analyse}
\alias{questionnaire_analyse}
\title{Analyse data from a standardised questionnaire}
\usage{
questionnaire_analyse(data, name = c("SUS", "Big5", "Big10"))
}
\arguments{
\item{data}{A data frame containing questionnaire responses.}

\item{name}{The name of the questionaire. Can be 'SUS', 'Big5', etc.}
}
\value{
The analysis of the questionnaire
}
\description{
\code{questionnaires_analyse} will simply print out a standard questionnaire,
and the citation for it.
}
\details{
The \code{data} must contain all the questions generated
by \code{\link{questionnaire_show}} as columns in a data frame. The
columns must be in the same order as in the questionnaire printed
by \code{\link{questionnaire_show}}.
}
\examples{
require(hcitools)
data(SUSresponses)
questionnaire_analyse(SUSresponses,'SUS')
}

