% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hcitools.R
\name{questionnaire_show}
\alias{questionnaire_show}
\title{Print a version of a standardised questionnaire}
\usage{
questionnaire_show(name = c("SUS", "Big5", "Big10"))
}
\arguments{
\item{name}{the name of the questionaire. Can be 'SUS', 'Big5', etc.}
}
\value{
Prints instructions, the questionnaire in plain text, and a citation.
}
\description{
\code{questionnaire_show} will simply print out a standard questionnaire,
and the citation for it
}
\examples{
questionnaire_show('SUS')
}

