% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hcitools.R
\name{questionnaires_show}
\alias{questionnaires_show}
\title{Print a version of a standardised questionnaire}
\usage{
questionnaires_show(name = c("SUS", "Big5", "Big10"))
}
\arguments{
\item{name}{the name of the questionaire. Can be 'SUS', 'Big5', etc.}
}
\value{
Prints instructions, the questionnaire in plain text, and a citation.
}
\description{
\code{questionnaires_show} will simply print out a standard questionnaire,
and the citation for it
}
\examples{
questionnaires_show('SUS')
}

