% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{SUSresponses}
\alias{SUSresponses}
\title{SUS dataset}
\format{Contains 45 survey responses (in rows) to the 10-point SUS questionnaire.}
\usage{
data(SUSresponses)
}
\description{
A sample dataset containing SUS questionnaire responses.
}
\examples{
data(SUSresponses)
questionnaires_analyse(SUSresponses,"SUS")
}
\keyword{datasets}

